Change partial download behavior
Instead of keeping track of download progress in downloading.json, Just keep the file name and url. If download needs to be resumed, a check is made to see if a partially downloaded file exists and download will resume from last downloaded byte up to total file size.

Also downloading file to a .part temporary name and renaming it when completed